added new entries to quiz.json, renamed hooks/ to api/, added new styles to delete button
