+++ public/quiz.json
@@ -143,11 +143,6 @@
                 "სავარაუდო პასუხი 4",
                 "სავარაუდო პასუხი 2"
               ]
-            },
-            {
-              "id": 3,
-              "type": "boolean",
-              "question": "დისკრეტული კითხვა"
             }
           ],
           "answers": [
@@ -161,10 +156,6 @@
                 2,
                 4
               ]
-            },
-            {
-              "id": 3,
-              "answer": true
             }
           ]
     },
@@ -199,6 +190,22 @@
               "id": 3,
               "type": "boolean",
               "question": "დისკრეტული კითხვა"
+            },
+            {
+              "id": 4,
+              "type": "boolean",
+              "question": "დისკრეტული კითხვა"
+            },
+            {
+              "id": 5,
+              "type": "multiple",
+              "question": "მრავალპასუხიანი კითხვა",
+              "options": [
+                "სავარაუდო პასუხი ^3",
+                "სავარაუდო პასუხი 1",
+                "სავარაუდო პასუხი ^4",
+                "სავარაუდო პასუხი 2"
+              ]
             }
           ],
           "answers": [
@@ -216,6 +223,17 @@
             {
               "id": 3,
               "answer": true
+            },
+            {
+              "id": 4,
+              "answer": false
+            },
+            {
+              "id": 5,
+              "answer": [
+                1,
+                3
+              ]
             }
           ]
     },
@@ -226,28 +244,6 @@
         "questions": [
             {
               "id": 1,
-              "type": "single",
-              "question": "ერთპასუხიანი კითხვა",
-              "options": [
-                "სავარაუდო პასუხი 1",
-                "სავარაუდო პასუხი 2",
-                "სავარაუდო პასუხი 3",
-                "სავარაუდო პასუხი 4"
-              ]
-            },
-            {
-              "id": 2,
-              "type": "multiple",
-              "question": "მრავალპასუხიანი კითხვა",
-              "options": [
-                "სავარაუდო პასუხი 3",
-                "სავარაუდო პასუხი 1",
-                "სავარაუდო პასუხი 4",
-                "სავარაუდო პასუხი 2"
-              ]
-            },
-            {
-              "id": 3,
               "type": "boolean",
               "question": "დისკრეტული კითხვა"
             }
@@ -255,17 +251,6 @@
           "answers": [
             {
               "id": 1,
-              "answer": 3
-            },
-            {
-              "id": 2,
-              "answer": [
-                2,
-                4
-              ]
-            },
-            {
-              "id": 3,
               "answer": true
             }
           ]
