+++ src/components/Base.css
@@ -69,6 +69,18 @@ h1 {
   font-size: 15px;
 }
 
+.btn-orange-slim {
+  background-color: #ff9500;
+  padding: 0px 15px;
+  border-radius: 6px;
+  border: none;
+  transition: 0.3s;
+  color: #fff;
+  height: 24px;
+  font-size: 14px;
+}
+
+
 .btn-orange-outline {
   background-color: #fff;
   padding: 0px 15px;
