+++ src/components/History/ContextMenu.jsx
@@ -7,8 +7,7 @@ const ContextMenu = () => {
     const handleListItemContextMenu = (e) => {
         e.preventDefault();
         
-        const identifier = e.target.parentElement.parentElement.parentElement.children[0].textContent.slice(9);
-        console.log(identifier)
+        const identifier = e.target.parentElement.children[0].textContent.slice(9);
         // მხოლოდ ერთი კონტექტ მენიუ არსებობდეს
         if (document.getElementById("context-menu")) {
             document.getElementById("context-menu").remove();
@@ -19,13 +18,14 @@ const ContextMenu = () => {
         let posW = e.clientX/window.innerWidth*100;
         let contextMenuAura = document.createElement("div");
         let contextMenu = document.createElement("div");
-        contextMenuAura.appendChild(contextMenu);
         contextMenuAura.id = "context-menu-aura";
-        let style = `width: 100%; height: 100%; background:rgba(0,0,0,0.2); position: fixed; top: 0; left: 0; z-index: 10;`
+        contextMenuAura.appendChild(contextMenu);
+        let style = `width: 100%; height: 100%; background:rgba(0,0,0,0); position: fixed; top: 0; left: 0; z-index: 10;`
         contextMenuAura.setAttribute("style", style);
         
         // მარცხენა ღილაკით წაშლა
         contextMenuAura.onclick = (e) => {
+            console.log(e.target.parentElement);
             if (e.target.parentElement.id!=="context-menu-aura"){
                 document.getElementById("context-menu-aura").remove();
             }
@@ -39,36 +39,41 @@ const ContextMenu = () => {
             }
         }
 
-        contextMenu.setAttribute("style", `position: absolute; z-index: 1; left: 0; top: 0;\
-            top: ${posH}%; left: ${posW}%; width: 100px;\
-            height: 50px; border: 3px solid #fff; background-color:${e.target.style.backgroundColor}`)
-        contextMenu.id = 'context-menu'
+        contextMenu.setAttribute("style", `position: fixed; z-index: 1; left: 0; top: 0;\
+            top: ${posH}%; left: ${posW}%; width: 100px; padding: 5px 3px;\
+            height: auto; border: 3px solid #fff; background-color: #ff9500`);
+        contextMenu.id = 'context-menu';
         contextMenu.textContent = e.target.textContent;
 
         let removeBtn = document.createElement("button");
-        removeBtn.textContent = "idk";
+        removeBtn.textContent = "yes";
         removeBtn.onclick = () => {
             if (identifier) {
                 let history = JSON.parse(localStorage.getItem("history"));
                 let temp = history.filter((el)=>el.date!==identifier);
                 localStorage.setItem("history", JSON.stringify(temp));
-                e.target.parentElement.parentElement.parentElement.remove();
+                e.target.parentElement.remove();
+            }
+        }
+
+        let cancelBtn = document.createElement("button");
+        cancelBtn.textContent = "cancel";
+        cancelBtn.onclick = (e) => {
+            e.preventDefault();
+            if (e.target.parentElement.id!=="context-menu-aura"){ 
+                document.getElementById("context-menu-aura").remove();
             }
         }
         
         document.getElementById("root").appendChild(contextMenuAura);
         contextMenu.appendChild(removeBtn);
+        contextMenu.appendChild(cancelBtn);
     }
     
     return (
-        <div>
-            <div className="list">
-                <div id="contextmenu-delete" className="list-item" 
-                    onContextMenu={handleListItemContextMenu} 
-                    style={{backgroundColor:"orange"}}>delete
-                </div>
-            </div>
-        </div>
+        <button className="btn-orange-slim" onContextMenu={handleListItemContextMenu} 
+            >delete
+        </button>
     )
 
 }
